+++ src/mpi/P2_mpi.cpp
@@ -171,7 +171,7 @@ T P2_mpi_master(T x, int64_t y, int threads)
 
     if (print_status())
     {
-      double percent = get_percent((double) low, (double) z);
+      double percent = get_percent(low, z);
       cout << "\rStatus: " << fixed << setprecision(get_status_precision(x))
            << percent << '%' << flush;
     }
